Embed css themes to simplify adding and testing new themes.
My idea here is that this change should simplify adding new themes for new contributors, and simplifies testing and adjusting themes for contributors and reviewers.
Currently we have to change 3 files (readme, settings.html, and the ever-growing style.css). I hope with this change we only need to drop in one new file and be done.

Note that during dev, rust-embed seems to read from disk, so "hot-reloading" changes to a theme file is possible.

I'm deep into "no idea what I'm doing here" with regard to some of the Rust code, so I'd be happy if anyone can show me how to make things cleaner.

List of TODOs: Remove duplicated ThemeAssets, resolve documentation in the readme, consider how to avoid assembling the style.css content every read (though this is cached, so maybe not so bad?).