(#38) redirect /:id to canonical URL for post (#58)
This is the first part of a possibly multi-stage fix toward #38. Specifically,
this implements redirection of `/:id` (a short-form URL to a post) to
the post's canonical URL. Ferrit issues a `HEAD /:id` to Reddit to get
the canonical URL, and on success will send an HTTP 302 to a client with
the canonical URL set in as the value of the `Location:` header.