Request CSS with explicit version.
base.html will now request with a query parameter `v=` whose value is
the current version of Libreddit. This will cause the browser to request
the stylesheet for a specific version of Libreddit, bypassing the cache.
A new version of Libreddit will cause the browser to fetch a new
stylesheet.

Resolves #622. Credit is due to GitHub user @chloekek for offering this
solution in the following post:
        https://github.com/libreddit/libreddit/issues/622#issuecomment-1315961742